[PATCH] kcapi.c CONFIG_MODULES=n build fix
With CONFIG_MODULES=n this file does not compile because the type of
module->owner is not known.

Gven that card->owner is probably a null pointer when this driver is
statically linked, best thing to do is to just not poke around inside
card->owner at all.

BKrev: 3f96a0bcJ9Rk6BnZoRDdwt64ukn4bg