Merge bk://kernel.bkbits.net/davem/net-2.5 into home.osdl.org:/home/torvalds/v2.5/linux
2003/10/27 08:37:48-08:00 kevcorry
[PATCH] Fix DM on top of raid

Force Device-Mapper to use PAGE_SIZE or smaller I/O when the underlying
device has a bvec_merge_fn routine registered.  This will fix the
situation of Device-Mapper submitting I/Os to RAID-0 that span the
RAID-0 chunk boundaries.

Joe is working on a better solution that actually honors the MD
merge_bvec_fn routine.  But this minimal change will fix the problem for
the time being.

2003/10/27 08:35:23-08:00 torvalds
Put the compiler barrier() on the right side of the preemption
enable on UP-PREEMPT.

Without this, the enable could "migrate" up into the critical
region (on SMP, the actual spinlock would act as an additional
barrier and PREEMPT was ok).

BKrev: 3f9d52fdvVMLnIdeZRhhwXHVFNuyLQ