Merge http://lia64.bkbits.net/to-linus-2.5 into home.osdl.org:/home/torvalds/v2.5/linux
2003/11/03 22:31:24-08:00 kochi
[PATCH] ia64: don't access per-CPU data of off-line CPUs

This patch prevents a crash that happens when per-CPU data is allocated
only for CPUs that are online.

2003/11/03 14:37:03-08:00 ajm
[PATCH] ia64: fix bug in SN2 sn_pci_map_sg that causes MCA

If sg->dma_address is set, we try to do a __pa() on a dma_address,
then, later, create a dma_addresss from a munged dma_address.  When
this bogus dma_address is used by the card, it results in MCAs.

BKrev: 3fa7cc3dwDculCwhpBjdhIU8aSWh8Q