[PATCH] Fix TSS limit on x86-64
The limit of the TSS segment was incorrectly set to a too big value
on x86-64. This lead to the CPU reading random memory behind the main
TSS when iopl was >0, but there was no ioperm bitmap set. This caused
random failures in port accesses in this state.

Set the correct limit.

BKrev: 3fad673ber4GuU7iWppydzNIyLntEQ