[PATCH] SMP signal latency fix
The code that sends a signal needs to "kick" the target process if it
runs on another CPU and wasn't woken up by the signal to let it know
that it has a new event.

Otherwise it might take a long time until the target actually notices
and acts on the signal.

BKrev: 3fad84fbDgjoiE4clsxi4TWA_uz69A