[PATCH] Warn about old EZD and DM disk remappers
Since 2.6.x by default does not remap the disk partitions any more, it
would be a good idea to warn when we see EZD or DM signatures.  That
informs the user to fix his setup (or pass in "hdx=remap" on the kernel
command line).

BKrev: 3fafc1dbAPkutcA5pWq30MQ2JI8xiQ