[PATCH] mpparse printk fix
From: Herbert Xu <herbert@gondor.apana.org.au>

The recent patch produces a message with no terminating newline on the
machine in question.  This is because one of the four bytes that you're
printing out is NUL.  The following patch avoids that problem.

BKrev: 3fbb26f3xWJN3d1ASmjO5nrKuAcXPQ