[PATCH] JFS: Avoid segfault when dirty inodes are written on readonly mount
This fixes an oops that can occur if JFS is used as the root filesystem.

Writes to a device node may cause a ->write_inode to be called during a
read-only mount.  JFS needs to check for NULL log in jfs_flush_journal.

BKrev: 3fc3cc7fMlaXUBErrUEu48crk6V2aw