Merge master.kernel.org:/home/davem/BK/net-2.5 into home.osdl.org:/home/torvalds/v2.5/linux
2003/12/03 12:59:12-08:00 mingo
[PATCH] Fix /proc access to dead thread group list oops

The pid_alive() check within the loop is incorrect.  If we are within
the tasklist lock and the thread group leader is valid then the thread
chain will be fully intact.

Instead, the check should be _outside_ the loop, since if the group
leader no longer exists, the whole list is gone and we must not try
to access it.

Move the check around, and add comment.

Bug-hunting and fix by Srivatsa Vaddagiri

BKrev: 3fce5748WwFTx_PDi6-0h9xC0NcZ7Q