Merge bk://gkernel.bkbits.net/net-drivers-2.5 into evo.osdl.org:/home/torvalds/v2.5/linux
2004/01/15 20:47:18-08:00 roland
[PATCH] fix pdeath_signal SMP locking

Obviously almost noone uses the pdeath_signal feature, since this has gone
unnoticed for quite some time.

This patch calls the function that does the right locking for the context
of this call (inside exit_notify).  The names of the signal.c entrypoints
are a little confusing.

BKrev: 40076d62opMjvhP7iVv0O-JY_CKUxQ