[PATCH] rq_for_each_bio fix
From: Xavier Bestel <xavier.bestel@free.fr>

Within the body of this macro we are accessing rq->bio, but `bio' is an arg
to the macro.  If someone uses this macro with some variable which is not
named `bio' it won't compile.

So use a more-likely-to-be-unique identifier for the macro.

BKrev: 400cb946Exb5XrEnaQmzdGDDSvSvHA